Add system for downloading and archiving zip files
This commit adds a system for downloading a zip file from a specified URL and archiving it on a remote server with a configurable retention period. The URL of the zip file to download is always the same. The downloaded file is then unzipped and compared with the file from the previous day. If they are different, a new archive is created with the current date and copied to the destination server.